refactor main orchestration
