+++ src/lib.rs
@@ -3,12 +3,14 @@
 //! Exposes core modules for use by binaries and tests.
 
 pub mod alerts;
+pub mod cli;
 pub mod config;
 pub mod engine;
 pub mod ioc;
 pub mod memory;
 pub mod models;
 pub mod normalizer;
+pub mod platform;
 pub mod reload;
 pub mod response;
 pub mod runtime;
