+++ src/runtime/mod.rs
@@ -1,3 +1,6 @@
 pub mod ioc;
 pub mod logging;
+mod orchestration;
 pub mod yara;
+
+pub use orchestration::run;
