+++ src/runtime/orchestration.rs
@@ -1,0 +1,1245 @@
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::cli::{Cli, Commands};
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::engine::{Engine, SigmaDetectionHandler};
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::ioc::IocEngine;
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::memory::MemoryScanConfig;
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::normalizer::Normalizer;
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::reload::DetectorStore;
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::response::ResponseEngine;
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::runtime::logging::{init_logging, log_startup_banner};
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::runtime::{ioc as runtime_ioc, yara as runtime_yara};
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::scanner::{YaraEventHandler, YaraMemoryJob};
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::sensor::{Platform, Sensor, SensorEvent, SensorEventRouter};
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::state::{ConnectionAggregator, DnsCache, ProcessCache, SidCache};
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use crate::{config, reload, scanner};
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use std::path::Path;
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use std::sync::Arc;
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+use tracing::{error, info, warn};
+
+#[cfg(windows)]
+use crate::sensor::windows::EtwSensor;
+#[cfg(windows)]
+use tokio::runtime::Builder;
+#[cfg(windows)]
+use tokio::sync::{mpsc, watch};
+
+#[cfg(target_os = "linux")]
+use crate::sensor::linux::EbpfSensor;
+#[cfg(target_os = "linux")]
+use tokio::sync::mpsc;
+
+#[cfg(target_os = "macos")]
+use crate::sensor::macos::EsfSensor;
+#[cfg(target_os = "macos")]
+use tokio::sync::mpsc;
+
+#[cfg(windows)]
+enum ShutdownMode {
+    Console,
+    Service(watch::Receiver<bool>),
+}
+
+#[cfg(windows)]
+pub fn run() -> anyhow::Result<()> {
+    if windows_service::service_dispatcher::start(
+        crate::platform::windows::SERVICE_NAME,
+        ffi_service_main,
+    )
+    .is_ok()
+    {
+        return Ok(());
+    }
+
+    let cli = Cli::parse_args();
+
+    match cli.command {
+        Some(Commands::Run { console }) => run_console(console, cli.log_level),
+        None => run_console(false, cli.log_level),
+        Some(Commands::Service { action }) => crate::platform::handle_service_command(action),
+    }
+}
+
+#[cfg(target_os = "linux")]
+pub fn run() -> anyhow::Result<()> {
+    let cli = Cli::parse_args();
+
+    match cli.command {
+        Some(Commands::Service { action }) => crate::platform::handle_service_command(action),
+        Some(Commands::Run { .. }) | None => run_linux(),
+    }
+}
+
+#[cfg(target_os = "linux")]
+fn run_linux() -> anyhow::Result<()> {
+    use tokio::runtime::Builder;
+    let runtime = Builder::new_multi_thread().enable_all().build()?;
+    runtime.block_on(run_linux_edr())
+}
+
+#[cfg(target_os = "macos")]
+pub fn run() -> anyhow::Result<()> {
+    let cli = Cli::parse_args();
+
+    match cli.command {
+        Some(Commands::Service { action }) => crate::platform::handle_service_command(action),
+        Some(Commands::Run { .. }) | None => run_macos(),
+    }
+}
+
+#[cfg(target_os = "macos")]
+fn run_macos() -> anyhow::Result<()> {
+    use tokio::runtime::Builder;
+    let runtime = Builder::new_multi_thread().enable_all().build()?;
+    runtime.block_on(run_macos_edr())
+}
+
+#[cfg(not(any(windows, target_os = "linux", target_os = "macos")))]
+pub fn run() -> anyhow::Result<()> {
+    Err(anyhow::anyhow!(
+        "This platform is not supported. Rustinel runs on Windows (ETW), Linux (eBPF), and macOS (ESF)."
+    ))
+}
+
+#[cfg(windows)]
+fn run_console(force_console: bool, log_level: Option<String>) -> anyhow::Result<()> {
+    let runtime = Builder::new_multi_thread().enable_all().build()?;
+    runtime.block_on(run_edr(ShutdownMode::Console, force_console, log_level))
+}
+
+#[cfg(windows)]
+extern "system" fn ffi_service_main(_args: u32, _raw_args: *mut *mut u16) {
+    if let Err(err) = service_main() {
+        eprintln!("Service error: {:?}", err);
+    }
+}
+
+#[cfg(windows)]
+fn service_main() -> anyhow::Result<()> {
+    use std::time::Duration;
+    use windows_service::service::{
+        ServiceControl, ServiceControlAccept, ServiceExitCode, ServiceState, ServiceStatus,
+        ServiceType,
+    };
+    use windows_service::service_control_handler::{self, ServiceControlHandlerResult};
+
+    let (shutdown_tx, shutdown_rx) = watch::channel(false);
+    let shutdown_tx = Arc::new(shutdown_tx);
+
+    let event_handler = move |control_event| -> ServiceControlHandlerResult {
+        match control_event {
+            ServiceControl::Stop | ServiceControl::Shutdown => {
+                let _ = shutdown_tx.send(true);
+                ServiceControlHandlerResult::NoError
+            }
+            ServiceControl::Interrogate => ServiceControlHandlerResult::NoError,
+            _ => ServiceControlHandlerResult::NotImplemented,
+        }
+    };
+
+    let status_handle =
+        service_control_handler::register(crate::platform::windows::SERVICE_NAME, event_handler)?;
+    let status_handle = Arc::new(status_handle);
+
+    status_handle.set_service_status(ServiceStatus {
+        service_type: ServiceType::OWN_PROCESS,
+        current_state: ServiceState::StartPending,
+        controls_accepted: ServiceControlAccept::empty(),
+        exit_code: ServiceExitCode::Win32(0),
+        checkpoint: 0,
+        wait_hint: Duration::from_secs(10),
+        process_id: None,
+    })?;
+
+    let runtime = Builder::new_multi_thread().enable_all().build()?;
+
+    status_handle.set_service_status(ServiceStatus {
+        service_type: ServiceType::OWN_PROCESS,
+        current_state: ServiceState::Running,
+        controls_accepted: ServiceControlAccept::STOP | ServiceControlAccept::SHUTDOWN,
+        exit_code: ServiceExitCode::Win32(0),
+        checkpoint: 0,
+        wait_hint: Duration::from_secs(0),
+        process_id: None,
+    })?;
+
+    let status_handle_for_stop = Arc::clone(&status_handle);
+    let mut stop_rx = shutdown_rx.clone();
+
+    let result = runtime.block_on(async move {
+        let stop_task = tokio::spawn(async move {
+            if stop_rx.changed().await.is_ok() {
+                let _ = status_handle_for_stop.set_service_status(ServiceStatus {
+                    service_type: ServiceType::OWN_PROCESS,
+                    current_state: ServiceState::StopPending,
+                    controls_accepted: ServiceControlAccept::empty(),
+                    exit_code: ServiceExitCode::Win32(0),
+                    checkpoint: 1,
+                    wait_hint: Duration::from_secs(10),
+                    process_id: None,
+                });
+            }
+        });
+
+        let run_result = run_edr(ShutdownMode::Service(shutdown_rx), false, None).await;
+        stop_task.abort();
+        let _ = stop_task.await;
+        run_result
+    });
+
+    let exit_code = if result.is_ok() {
+        ServiceExitCode::Win32(0)
+    } else {
+        ServiceExitCode::ServiceSpecific(1)
+    };
+
+    status_handle.set_service_status(ServiceStatus {
+        service_type: ServiceType::OWN_PROCESS,
+        current_state: ServiceState::Stopped,
+        controls_accepted: ServiceControlAccept::empty(),
+        exit_code,
+        checkpoint: 0,
+        wait_hint: Duration::from_secs(0),
+        process_id: None,
+    })?;
+
+    result
+}
+
+#[cfg(windows)]
+fn spawn_shutdown_handler(
+    shutdown_mode: ShutdownMode,
+    sensor: Arc<EtwSensor>,
+) -> tokio::task::JoinHandle<()> {
+    tokio::spawn(async move {
+        match shutdown_mode {
+            ShutdownMode::Console => match tokio::signal::ctrl_c().await {
+                Ok(()) => {
+                    info!("Received Ctrl+C signal");
+                    sensor.shutdown();
+                }
+                Err(err) => {
+                    error!("Failed to listen for Ctrl+C: {}", err);
+                }
+            },
+            ShutdownMode::Service(mut shutdown_rx) => {
+                if shutdown_rx.changed().await.is_ok() {
+                    info!("Received service stop signal");
+                } else {
+                    warn!("Service shutdown channel dropped");
+                }
+                sensor.shutdown();
+            }
+        }
+    })
+}
+
+#[cfg(windows)]
+async fn run_edr(
+    shutdown_mode: ShutdownMode,
+    force_console: bool,
+    log_level_override: Option<String>,
+) -> anyhow::Result<()> {
+    // 1. Load Configuration
+    let mut cfg = match config::AppConfig::new() {
+        Ok(cfg) => cfg,
+        Err(err) => {
+            eprintln!("Failed to load configuration: {}", err);
+            eprintln!("Hint: check config.toml and EDR__* environment overrides.");
+            return Err(anyhow::anyhow!("Failed to load configuration: {}", err));
+        }
+    };
+    if force_console {
+        cfg.logging.console_output = true;
+    }
+    if let Some(level) = log_level_override {
+        if !level.trim().is_empty() {
+            cfg.logging.level = level;
+        }
+    }
+
+    // 2. Initialize Logging (CRITICAL: Store guards to keep file writing alive)
+    let (app_guard, alert_guard, alert_sink) = init_logging(&cfg);
+    let _guards = (app_guard, alert_guard);
+
+    log_startup_banner("Windows ETW");
+
+    // 2.1 Initialize Active Response Engine (optional)
+    let (response_engine, response_worker_handle) = ResponseEngine::new(&cfg.response);
+    info!(
+        target: "rustinel",
+        logs_dir = ?cfg.logging.directory,
+        alerts_dir = ?cfg.alerts.directory,
+        "Agent started with dual-pipeline logging"
+    );
+
+    // Verify running with appropriate privileges
+    #[cfg(windows)]
+    {
+        use windows::Win32::Foundation::HANDLE;
+        use windows::Win32::Security::{
+            GetTokenInformation, TokenElevation, TOKEN_ELEVATION, TOKEN_QUERY,
+        };
+        use windows::Win32::System::Threading::{GetCurrentProcess, OpenProcessToken};
+
+        unsafe {
+            let mut token = HANDLE::default();
+            if OpenProcessToken(GetCurrentProcess(), TOKEN_QUERY, &mut token).is_ok() {
+                let mut elevation = TOKEN_ELEVATION::default();
+                let mut return_length = 0u32;
+
+                if GetTokenInformation(
+                    token,
+                    TokenElevation,
+                    Some(&mut elevation as *mut _ as *mut _),
+                    std::mem::size_of::<TOKEN_ELEVATION>() as u32,
+                    &mut return_length,
+                )
+                .is_ok()
+                {
+                    if elevation.TokenIsElevated == 0 {
+                        error!("❌ ERROR: This application requires Administrator privileges!");
+                        error!("   Please run as Administrator to access ETW providers.");
+                        return Err(anyhow::anyhow!(
+                            "Insufficient privileges - Administrator access required"
+                        ));
+                    } else {
+                        info!("✓ Running with Administrator privileges");
+                    }
+                }
+            }
+        }
+    }
+
+    // Initialize modules
+    info!("Initializing modules...");
+
+    // Initialize Process Cache and perform cold start snapshot
+    info!("Initializing Process Cache...");
+    let process_cache = Arc::new(ProcessCache::new());
+    let sid_cache = Arc::new(SidCache::new());
+    let dns_cache = Arc::new(DnsCache::new());
+    let connection_aggregator = Arc::new(ConnectionAggregator::with_limits(
+        cfg.network.aggregation_max_entries,
+        cfg.network.aggregation_interval_buffer_size,
+    ));
+
+    // Snapshot existing processes using Windows API (handles cold start problem)
+    #[cfg(windows)]
+    {
+        match crate::platform::windows::snapshot_processes(&process_cache) {
+            Ok(count) => {
+                info!(
+                    "✓ Process Cache initialized with {} existing processes",
+                    count
+                );
+            }
+            Err(e) => {
+                warn!(
+                    "Failed to snapshot processes: {}. Cache will populate from ETW events.",
+                    e
+                );
+            }
+        }
+    }
+
+    #[cfg(not(windows))]
+    {
+        info!("Process snapshot not available on non-Windows platforms");
+    }
+
+    let sensor = Arc::new(EtwSensor::new());
+
+    // Initialize Sigma engine
+    let mut sigma_engine = Engine::new_for_platform_with_logging_level_and_match_debug(
+        Platform::Windows,
+        &cfg.logging.level,
+        cfg.alerts.match_debug,
+    );
+
+    if cfg.scanner.sigma_enabled {
+        info!(
+            target: "rustinel",
+            rules_path = ?cfg.scanner.sigma_rules_path,
+            "Loading Sigma rules"
+        );
+
+        if let Err(e) = sigma_engine.load_rules(&cfg.scanner.sigma_rules_path) {
+            warn!(target: "rustinel", error = %e, "Failed to load Sigma rules");
+        } else {
+            let stats = sigma_engine.stats();
+            info!(
+                target: "rustinel",
+                total_rules = stats.total_rules,
+                skipped_deferred_rules = stats.skipped_deferred_rules,
+                skipped_unknown_logsource_rules = stats.skipped_unknown_logsource_rules,
+                skipped_product_rules = stats.skipped_product_rules,
+                inactive_collector_rules = stats.inactive_collector_rules,
+                "Sigma Engine initialized"
+            );
+            for (logsource, count) in stats.rules_by_logsource {
+                info!(target: "rustinel", logsource = %logsource, count = count, "Sigma rules loaded");
+            }
+        }
+    } else {
+        info!(target: "rustinel", "Sigma detection disabled by configuration");
+    }
+    let sigma_engine = Arc::new(sigma_engine);
+
+    // Initialize YARA Scanner
+    let yara_scanner = if cfg.scanner.yara_enabled {
+        info!(
+            target: "rustinel",
+            rules_path = ?cfg.scanner.yara_rules_path,
+            "Initializing YARA Scanner"
+        );
+
+        match scanner::Scanner::new(&cfg.scanner.yara_rules_path) {
+            Ok(s) => {
+                info!(target: "rustinel", "YARA Scanner initialized successfully");
+                Arc::new(s)
+            }
+            Err(e) => {
+                warn!(target: "rustinel", error = %e, "Failed to load YARA rules. YARA scanning disabled.");
+                // Create an empty scanner so we don't crash
+                Arc::new(
+                    scanner::Scanner::new(Path::new("."))
+                        .expect("Failed to create empty YARA scanner"),
+                )
+            }
+        }
+    } else {
+        info!(target: "rustinel", "YARA scanning disabled by configuration");
+        Arc::new(
+            scanner::Scanner::new(Path::new(".")).expect("Failed to create empty YARA scanner"),
+        )
+    };
+
+    let yara_allowlist_paths =
+        scanner::normalize_allowlist_paths(&cfg.scanner.yara_allowlist_paths);
+    if !yara_allowlist_paths.is_empty() {
+        info!(
+            target: "rustinel",
+            count = yara_allowlist_paths.len(),
+            "YARA allowlist paths loaded (files under these paths will NOT be scanned)"
+        );
+    }
+
+    // Create background worker channel for YARA scanning
+    // Buffer = 1000 items. If 1000 processes start instantly, we drop events rather than blocking.
+    let (tx, rx) = mpsc::channel::<(String, u32)>(1000);
+
+    // Create optional YARA memory scanning channel.
+    let (yara_memory_tx, yara_memory_rx) =
+        if cfg.scanner.yara_enabled && cfg.scanner.yara_memory_enabled {
+            let capacity = cfg.scanner.yara_memory_queue_capacity.max(1);
+            let (tx, rx) = mpsc::channel::<YaraMemoryJob>(capacity);
+            (Some(tx), Some(rx))
+        } else {
+            (None, None)
+        };
+
+    // Initialize IOC engine
+    let ioc_engine = Arc::new(IocEngine::load(&cfg.ioc));
+    if ioc_engine.is_enabled() {
+        let stats = ioc_engine.stats();
+        info!(
+            target: "rustinel",
+            md5 = stats.md5,
+            sha1 = stats.sha1,
+            sha256 = stats.sha256,
+            ip = stats.ip,
+            cidr = stats.cidr,
+            domain_exact = stats.domain_exact,
+            domain_suffix = stats.domain_suffix,
+            path_regex = stats.path_regex,
+            "IOC engine initialized"
+        );
+    } else {
+        info!(target: "rustinel", "IOC detection disabled by configuration");
+    }
+
+    let detectors = DetectorStore::new(
+        Arc::clone(&sigma_engine),
+        Arc::clone(&yara_scanner),
+        Arc::clone(&ioc_engine),
+    );
+
+    let mut reload_poller_handle = None;
+    let mut reload_worker_handle = None;
+    let mut reload_tx = None;
+    if cfg.reload.enabled {
+        let (tx, rx) = mpsc::unbounded_channel();
+
+        reload_worker_handle = Some(reload::spawn_reload_worker(
+            Arc::clone(&detectors),
+            cfg.scanner.clone(),
+            cfg.ioc.clone(),
+            cfg.reload.clone(),
+            cfg.logging.level.clone(),
+            cfg.alerts.match_debug,
+            rx,
+        ));
+
+        reload_poller_handle = Some(reload::spawn_reload_poller(
+            cfg.scanner.clone(),
+            cfg.ioc.clone(),
+            cfg.reload.clone(),
+            tx.clone(),
+        ));
+        reload_tx = Some(tx);
+    } else {
+        info!(target: "reload", "Hot-reload disabled by configuration");
+    }
+
+    let yara_worker_handle = runtime_yara::spawn_yara_file_worker(
+        Arc::clone(&detectors),
+        alert_sink.clone(),
+        response_engine.clone(),
+        cfg.alerts.match_debug,
+        rx,
+        yara_allowlist_paths.clone(),
+        Platform::Windows,
+        "etw",
+    );
+
+    // Spawn optional YARA memory scanning worker.
+    let yara_memory_worker_handle = if let Some(mem_rx) = yara_memory_rx {
+        let mem_cfg = MemoryScanConfig {
+            max_process_bytes: (cfg.scanner.yara_memory_max_process_mb * 1024 * 1024) as usize,
+            max_region_bytes: (cfg.scanner.yara_memory_max_region_mb * 1024 * 1024) as usize,
+            include_private: cfg.scanner.yara_memory_include_private,
+            include_image: cfg.scanner.yara_memory_include_image,
+            include_mapped: cfg.scanner.yara_memory_include_mapped,
+            delay_ms: cfg.scanner.yara_memory_delay_ms,
+        };
+        Some(runtime_yara::spawn_yara_memory_worker(
+            Arc::clone(&detectors),
+            alert_sink.clone(),
+            response_engine.clone(),
+            mem_cfg,
+            cfg.alerts.match_debug,
+            mem_rx,
+            Platform::Windows,
+            "yara-memory",
+        ))
+    } else {
+        None
+    };
+
+    // Create background worker channel for IOC hashing (process start only)
+    // Uses spawn_blocking to avoid starving the tokio async thread pool with
+    // CPU-bound crypto work and synchronous file I/O.
+    let (ioc_hash_tx, mut ioc_hash_worker_handle) = if ioc_engine.is_enabled() {
+        let (hash_tx, hash_rx) = mpsc::channel::<(String, u32)>(1000);
+        let hash_worker_handle = runtime_ioc::spawn_ioc_hash_worker(
+            Arc::clone(&detectors),
+            alert_sink.clone(),
+            response_engine.clone(),
+            hash_rx,
+            Platform::Windows,
+            "etw",
+        );
+
+        (Some(hash_tx), Some(hash_worker_handle))
+    } else {
+        (None, None)
+    };
+
+    // Initialize normalizer with process cache and connection aggregator
+    let normalizer = Arc::new(Normalizer::new(
+        Arc::clone(&process_cache),
+        Arc::clone(&sid_cache),
+        Arc::clone(&dns_cache),
+        Arc::clone(&connection_aggregator),
+        cfg.network.aggregation_enabled,
+    ));
+
+    info!("✓ ETW sensor initialized");
+    info!("✓ Normalizer initialized");
+
+    // Create Sigma detection handler
+    let sigma_handler = SigmaDetectionHandler {
+        normalizer: Arc::clone(&normalizer),
+        detectors: Arc::clone(&detectors),
+        ioc_hash_tx,
+        alert_sink: alert_sink.clone(),
+        response_engine: response_engine.clone(),
+    };
+
+    // Create YARA event handler
+    let yara_handler = YaraEventHandler {
+        tx,
+        memory_tx: yara_memory_tx,
+        allowlist_paths: yara_allowlist_paths,
+    };
+
+    // Setup shared SensorEventRouter (mutable)
+    let mut router_inner = SensorEventRouter::new();
+    router_inner.register_handler(Box::new(sigma_handler));
+    router_inner.register_handler(Box::new(yara_handler));
+
+    // Freeze router (immutable/shared)
+    let router = Arc::new(router_inner);
+
+    info!("✓ Event Router initialized");
+    info!("✓ Event handlers registered");
+
+    // Setup graceful shutdown handler
+    let shutdown_handler = spawn_shutdown_handler(shutdown_mode, Arc::clone(&sensor));
+
+    info!("✓ Signal handlers configured");
+    info!("");
+    info!("Starting ETW trace session...");
+    info!("Press Ctrl+C to stop gracefully");
+    info!("");
+
+    // Start shared sensor event pipeline
+    let (sensor_tx, mut sensor_rx) = mpsc::channel::<SensorEvent>(8192);
+    let router_clone = Arc::clone(&router);
+    let sensor_worker_handle = tokio::task::spawn_blocking(move || {
+        info!(target: "sensor", "Sensor event worker thread started");
+        while let Some(event) = sensor_rx.blocking_recv() {
+            router_clone.route_event(&event);
+        }
+        info!(target: "sensor", "Sensor event worker thread shutting down");
+    });
+
+    let sensor_clone = Arc::clone(&sensor);
+
+    // We make trace_handle mutable so we can await it.
+    let mut trace_handle = tokio::task::spawn_blocking(move || {
+        if let Err(e) = sensor_clone.start(sensor_tx) {
+            error!("ETW sensor error: {}", e);
+        }
+    });
+
+    // Wait for either shutdown signal or trace completion.
+    tokio::select! {
+        _ = shutdown_handler => {
+            info!("Shutdown signal received, waiting for ETW session to close...");
+            match trace_handle.await {
+                Ok(_) => info!("ETW sensor thread finished"),
+                Err(e) => error!("Failed to join ETW sensor thread: {}", e),
+            }
+        }
+        // CRITICAL: If trace finishes unexpectedly, the ETW sensor died.
+        // This means the agent is "blind" - still running but not collecting events.
+        result = &mut trace_handle => {
+            if sensor.is_shutdown() {
+                info!("ETW sensor thread finished after shutdown request");
+            } else {
+                error!("🚨 CRITICAL: ETW sensor thread died unexpectedly!");
+                match result {
+                    Ok(_) => {
+                        error!("Trace stopped without panic (unexpected normal termination)");
+                        error!("This indicates the ETW session closed unexpectedly");
+                    }
+                    Err(join_err) => {
+                        if join_err.is_panic() {
+                            error!("🔥 PANIC: Trace thread PANICKED!");
+                            // Try to extract panic message (into_panic consumes join_err)
+                            let panic_info = join_err.into_panic();
+                            if let Some(panic_msg) = panic_info.downcast_ref::<&str>() {
+                                error!("Panic message: {}", panic_msg);
+                            } else if let Some(panic_msg) = panic_info.downcast_ref::<String>() {
+                                error!("Panic message: {}", panic_msg);
+                            } else {
+                                error!("Panic message: <unable to extract>");
+                            }
+                        } else {
+                            error!("Trace thread cancelled/failed: {}", join_err);
+                        }
+                    }
+                }
+                // Force exit so Service Manager/Watchdog restarts the agent
+                // Without this, the agent appears "Online" but is blind to events
+                error!("Forcing process exit to trigger restart...");
+                std::process::exit(1);
+            }
+        }
+    }
+
+    // Common teardown (not reached if exit(1) above)
+    match sensor_worker_handle.await {
+        Ok(_) => info!("Sensor event worker thread finished"),
+        Err(e) => error!("Failed to join sensor event worker thread: {}", e),
+    }
+
+    drop(router);
+    drop(response_engine);
+    info!("Signaling YARA worker to shut down...");
+
+    match yara_worker_handle.await {
+        Ok(_) => info!("YARA worker thread finished"),
+        Err(e) => error!("Failed to join YARA worker thread: {}", e),
+    }
+
+    if let Some(handle) = yara_memory_worker_handle {
+        match handle.await {
+            Ok(_) => info!("YARA memory worker thread finished"),
+            Err(e) => error!("Failed to join YARA memory worker thread: {}", e),
+        }
+    }
+
+    if let Some(handle) = ioc_hash_worker_handle.take() {
+        info!("Signaling IOC hash worker to shut down...");
+        match handle.await {
+            Ok(_) => info!("IOC hash worker thread finished"),
+            Err(e) => error!("Failed to join IOC hash worker thread: {}", e),
+        }
+    }
+
+    if let Some(handle) = reload_poller_handle.take() {
+        info!("Signaling hot-reload poller to shut down...");
+        handle.abort();
+        let _ = handle.await;
+        info!("Hot-reload poller thread finished");
+    }
+    drop(reload_tx.take());
+    if let Some(handle) = reload_worker_handle.take() {
+        info!("Signaling hot-reload worker to shut down...");
+        match handle.await {
+            Ok(_) => info!("Hot-reload worker thread finished"),
+            Err(e) => error!("Failed to join hot-reload worker thread: {}", e),
+        }
+    }
+
+    info!("Signaling response worker to shut down...");
+    match response_worker_handle.await {
+        Ok(_) => info!("Response worker thread finished"),
+        Err(e) => error!("Failed to join response worker thread: {}", e),
+    }
+
+    info!("");
+    info!("╔═══════════════════════════════════════════════════╗");
+    info!("║           Shutdown Complete                       ║");
+    info!("║        Thank you for using Rustinel!              ║");
+    info!("╚═══════════════════════════════════════════════════╝");
+
+    Ok(())
+}
+
+// ── Linux eBPF EDR entry point ────────────────────────────────────────────────
+
+/// Linux eBPF EDR main loop. Mirrors `run_edr` but replaces ETW with the
+/// eBPF sensor and omits Windows-only subsystems.
+#[cfg(target_os = "linux")]
+async fn run_linux_edr() -> anyhow::Result<()> {
+    // 1. Configuration
+    let cfg = match config::AppConfig::new() {
+        Ok(cfg) => cfg,
+        Err(err) => {
+            eprintln!("Failed to load configuration: {}", err);
+            return Err(anyhow::anyhow!("configuration error: {}", err));
+        }
+    };
+
+    // 2. Logging
+    let (app_guard, alert_guard, alert_sink) = init_logging(&cfg);
+    let _guards = (app_guard, alert_guard);
+
+    log_startup_banner("Linux eBPF");
+
+    // 3. Shared state
+    let process_cache = Arc::new(ProcessCache::new());
+    let sid_cache = Arc::new(SidCache::new()); // no-op on Linux; kept for Normalizer compat
+    let dns_cache = Arc::new(DnsCache::new());
+    let connection_aggregator = Arc::new(ConnectionAggregator::with_limits(
+        cfg.network.aggregation_max_entries,
+        cfg.network.aggregation_interval_buffer_size,
+    ));
+
+    // 4. Active response engine
+    let (response_engine, response_worker_handle) = ResponseEngine::new(&cfg.response);
+
+    // 5. Sigma engine
+    let mut sigma_engine = Engine::new_for_platform_with_logging_level_and_match_debug(
+        Platform::Linux,
+        &cfg.logging.level,
+        cfg.alerts.match_debug,
+    );
+
+    if cfg.scanner.sigma_enabled {
+        info!(rules_path = ?cfg.scanner.sigma_rules_path, "Loading Sigma rules");
+        if let Err(e) = sigma_engine.load_rules(&cfg.scanner.sigma_rules_path) {
+            warn!(error = %e, "Failed to load Sigma rules");
+        } else {
+            let stats = sigma_engine.stats();
+            info!(
+                total_rules = stats.total_rules,
+                skipped_deferred_rules = stats.skipped_deferred_rules,
+                skipped_unknown_logsource_rules = stats.skipped_unknown_logsource_rules,
+                skipped_product_rules = stats.skipped_product_rules,
+                inactive_collector_rules = stats.inactive_collector_rules,
+                "Sigma engine initialized"
+            );
+            for (logsource, count) in stats.rules_by_logsource {
+                info!(logsource = %logsource, count, "Sigma rules loaded");
+            }
+        }
+    }
+    let sigma_engine = Arc::new(sigma_engine);
+
+    // 6. YARA scanner
+    let yara_scanner = if cfg.scanner.yara_enabled {
+        match scanner::Scanner::new(&cfg.scanner.yara_rules_path) {
+            Ok(s) => {
+                info!("YARA scanner initialized");
+                Arc::new(s)
+            }
+            Err(e) => {
+                warn!(error = %e, "Failed to load YARA rules; YARA scanning disabled");
+                Arc::new(scanner::Scanner::new(Path::new(".")).expect("empty YARA scanner"))
+            }
+        }
+    } else {
+        Arc::new(scanner::Scanner::new(Path::new(".")).expect("empty YARA scanner"))
+    };
+
+    let yara_allowlist_paths =
+        scanner::normalize_allowlist_paths(&cfg.scanner.yara_allowlist_paths);
+
+    // 7. IOC engine
+    let ioc_engine = Arc::new(IocEngine::load(&cfg.ioc));
+
+    // 8. Detector store + hot-reload
+    let detectors = DetectorStore::new(
+        Arc::clone(&sigma_engine),
+        Arc::clone(&yara_scanner),
+        Arc::clone(&ioc_engine),
+    );
+
+    let mut reload_poller_handle = None;
+    let mut reload_worker_handle = None;
+    let mut reload_tx = None;
+    if cfg.reload.enabled {
+        let (tx, rx) = mpsc::unbounded_channel();
+        reload_worker_handle = Some(reload::spawn_reload_worker(
+            Arc::clone(&detectors),
+            cfg.scanner.clone(),
+            cfg.ioc.clone(),
+            cfg.reload.clone(),
+            cfg.logging.level.clone(),
+            cfg.alerts.match_debug,
+            rx,
+        ));
+        reload_poller_handle = Some(reload::spawn_reload_poller(
+            cfg.scanner.clone(),
+            cfg.ioc.clone(),
+            cfg.reload.clone(),
+            tx.clone(),
+        ));
+        reload_tx = Some(tx);
+    }
+
+    // 9. YARA background worker
+    let (yara_tx, yara_rx) = mpsc::channel::<(String, u32)>(1000);
+
+    let (yara_memory_tx, yara_memory_rx) =
+        if cfg.scanner.yara_enabled && cfg.scanner.yara_memory_enabled {
+            let capacity = cfg.scanner.yara_memory_queue_capacity.max(1);
+            let (tx, rx) = mpsc::channel::<YaraMemoryJob>(capacity);
+            (Some(tx), Some(rx))
+        } else {
+            (None, None)
+        };
+    let yara_worker_handle = runtime_yara::spawn_yara_file_worker(
+        Arc::clone(&detectors),
+        alert_sink.clone(),
+        response_engine.clone(),
+        cfg.alerts.match_debug,
+        yara_rx,
+        yara_allowlist_paths.clone(),
+        Platform::Linux,
+        "ebpf",
+    );
+
+    // Spawn optional YARA memory scanning worker (Linux).
+    let yara_memory_worker_handle = if let Some(mem_rx) = yara_memory_rx {
+        let mem_cfg = MemoryScanConfig {
+            max_process_bytes: (cfg.scanner.yara_memory_max_process_mb * 1024 * 1024) as usize,
+            max_region_bytes: (cfg.scanner.yara_memory_max_region_mb * 1024 * 1024) as usize,
+            include_private: cfg.scanner.yara_memory_include_private,
+            include_image: cfg.scanner.yara_memory_include_image,
+            include_mapped: cfg.scanner.yara_memory_include_mapped,
+            delay_ms: cfg.scanner.yara_memory_delay_ms,
+        };
+        Some(runtime_yara::spawn_yara_memory_worker(
+            Arc::clone(&detectors),
+            alert_sink.clone(),
+            response_engine.clone(),
+            mem_cfg,
+            cfg.alerts.match_debug,
+            mem_rx,
+            Platform::Linux,
+            "yara-memory",
+        ))
+    } else {
+        None
+    };
+
+    // 10. IOC hash background worker
+    let (ioc_hash_tx, mut ioc_hash_worker_handle) = if ioc_engine.is_enabled() {
+        let (hash_tx, hash_rx) = mpsc::channel::<(String, u32)>(1000);
+        let handle = runtime_ioc::spawn_ioc_hash_worker(
+            Arc::clone(&detectors),
+            alert_sink.clone(),
+            response_engine.clone(),
+            hash_rx,
+            Platform::Linux,
+            "ebpf",
+        );
+        (Some(hash_tx), Some(handle))
+    } else {
+        (None, None)
+    };
+
+    // 11. Normalizer
+    let normalizer = Arc::new(Normalizer::new(
+        Arc::clone(&process_cache),
+        Arc::clone(&sid_cache),
+        Arc::clone(&dns_cache),
+        Arc::clone(&connection_aggregator),
+        cfg.network.aggregation_enabled,
+    ));
+
+    // 12. Detection handlers + router
+    let sigma_handler = SigmaDetectionHandler {
+        normalizer: Arc::clone(&normalizer),
+        detectors: Arc::clone(&detectors),
+        ioc_hash_tx,
+        alert_sink: alert_sink.clone(),
+        response_engine: response_engine.clone(),
+    };
+    let yara_handler = YaraEventHandler {
+        tx: yara_tx,
+        memory_tx: yara_memory_tx,
+        allowlist_paths: yara_allowlist_paths,
+    };
+    let mut router_inner = SensorEventRouter::new();
+    router_inner.register_handler(Box::new(sigma_handler));
+    router_inner.register_handler(Box::new(yara_handler));
+    let router = Arc::new(router_inner);
+
+    // 13. eBPF sensor
+    let sensor = Arc::new(EbpfSensor::new());
+
+    info!("Starting eBPF sensor...");
+    info!("Press Ctrl+C to stop gracefully");
+
+    let (sensor_tx, mut sensor_rx) = mpsc::channel::<SensorEvent>(8192);
+    let router_for_worker = Arc::clone(&router);
+    let sensor_worker_handle = tokio::task::spawn_blocking(move || {
+        while let Some(event) = sensor_rx.blocking_recv() {
+            router_for_worker.route_event(&event);
+        }
+    });
+
+    let sensor_clone = Arc::clone(&sensor);
+    if let Err(e) = sensor_clone.start(sensor_tx) {
+        error!("eBPF sensor failed to start: {:#}", e);
+        return Err(e);
+    }
+
+    // 14. Wait for Ctrl+C
+    match tokio::signal::ctrl_c().await {
+        Ok(()) => info!("Received Ctrl+C, shutting down"),
+        Err(e) => error!("Failed to listen for Ctrl+C: {}", e),
+    }
+    sensor.shutdown();
+
+    // Drain workers
+    drop(router);
+    drop(response_engine);
+    let _ = sensor_worker_handle.await;
+    let _ = yara_worker_handle.await;
+    if let Some(h) = yara_memory_worker_handle {
+        let _ = h.await;
+    }
+    if let Some(h) = ioc_hash_worker_handle.take() {
+        let _ = h.await;
+    }
+    if let Some(h) = reload_poller_handle.take() {
+        h.abort();
+        let _ = h.await;
+    }
+    drop(reload_tx.take());
+    if let Some(h) = reload_worker_handle.take() {
+        let _ = h.await;
+    }
+    let _ = response_worker_handle.await;
+
+    info!("Shutdown complete");
+    Ok(())
+}
+
+// ── macOS EDR entry point ─────────────────────────────────────────────────────
+
+/// macOS EDR main loop. Mirrors `run_linux_edr` but replaces the eBPF sensor
+/// with the macOS sensor. In Phase 0 this is a no-op placeholder sensor so the
+/// shared pipeline can be exercised end to end; the Endpoint Security and
+/// /dev/bpf sources land in later phases.
+#[cfg(target_os = "macos")]
+async fn run_macos_edr() -> anyhow::Result<()> {
+    // 1. Configuration
+    let cfg = match config::AppConfig::new() {
+        Ok(cfg) => cfg,
+        Err(err) => {
+            eprintln!("Failed to load configuration: {}", err);
+            return Err(anyhow::anyhow!("configuration error: {}", err));
+        }
+    };
+
+    // 2. Logging
+    let (app_guard, alert_guard, alert_sink) = init_logging(&cfg);
+    let _guards = (app_guard, alert_guard);
+
+    log_startup_banner("macOS ESF");
+
+    // 3. Shared state
+    let process_cache = Arc::new(ProcessCache::new());
+    let sid_cache = Arc::new(SidCache::new()); // no-op on macOS; kept for Normalizer compat
+    let dns_cache = Arc::new(DnsCache::new());
+    let connection_aggregator = Arc::new(ConnectionAggregator::with_limits(
+        cfg.network.aggregation_max_entries,
+        cfg.network.aggregation_interval_buffer_size,
+    ));
+
+    // 4. Active response engine
+    let (response_engine, response_worker_handle) = ResponseEngine::new(&cfg.response);
+
+    // 5. Sigma engine
+    let mut sigma_engine = Engine::new_for_platform_with_logging_level_and_match_debug(
+        Platform::MacOS,
+        &cfg.logging.level,
+        cfg.alerts.match_debug,
+    );
+
+    if cfg.scanner.sigma_enabled {
+        info!(rules_path = ?cfg.scanner.sigma_rules_path, "Loading Sigma rules");
+        if let Err(e) = sigma_engine.load_rules(&cfg.scanner.sigma_rules_path) {
+            warn!(error = %e, "Failed to load Sigma rules");
+        } else {
+            let stats = sigma_engine.stats();
+            info!(
+                total_rules = stats.total_rules,
+                skipped_deferred_rules = stats.skipped_deferred_rules,
+                skipped_unknown_logsource_rules = stats.skipped_unknown_logsource_rules,
+                skipped_product_rules = stats.skipped_product_rules,
+                inactive_collector_rules = stats.inactive_collector_rules,
+                "Sigma engine initialized"
+            );
+            for (logsource, count) in stats.rules_by_logsource {
+                info!(logsource = %logsource, count, "Sigma rules loaded");
+            }
+        }
+    }
+    let sigma_engine = Arc::new(sigma_engine);
+
+    // 6. YARA scanner
+    let yara_scanner = if cfg.scanner.yara_enabled {
+        match scanner::Scanner::new(&cfg.scanner.yara_rules_path) {
+            Ok(s) => {
+                info!("YARA scanner initialized");
+                Arc::new(s)
+            }
+            Err(e) => {
+                warn!(error = %e, "Failed to load YARA rules; YARA scanning disabled");
+                Arc::new(scanner::Scanner::new(Path::new(".")).expect("empty YARA scanner"))
+            }
+        }
+    } else {
+        Arc::new(scanner::Scanner::new(Path::new(".")).expect("empty YARA scanner"))
+    };
+
+    let yara_allowlist_paths =
+        scanner::normalize_allowlist_paths(&cfg.scanner.yara_allowlist_paths);
+
+    // 7. IOC engine
+    let ioc_engine = Arc::new(IocEngine::load(&cfg.ioc));
+
+    // 8. Detector store + hot-reload
+    let detectors = DetectorStore::new(
+        Arc::clone(&sigma_engine),
+        Arc::clone(&yara_scanner),
+        Arc::clone(&ioc_engine),
+    );
+
+    let mut reload_poller_handle = None;
+    let mut reload_worker_handle = None;
+    let mut reload_tx = None;
+    if cfg.reload.enabled {
+        let (tx, rx) = mpsc::unbounded_channel();
+        reload_worker_handle = Some(reload::spawn_reload_worker(
+            Arc::clone(&detectors),
+            cfg.scanner.clone(),
+            cfg.ioc.clone(),
+            cfg.reload.clone(),
+            cfg.logging.level.clone(),
+            cfg.alerts.match_debug,
+            rx,
+        ));
+        reload_poller_handle = Some(reload::spawn_reload_poller(
+            cfg.scanner.clone(),
+            cfg.ioc.clone(),
+            cfg.reload.clone(),
+            tx.clone(),
+        ));
+        reload_tx = Some(tx);
+    }
+
+    // 9. YARA background worker
+    let (yara_tx, yara_rx) = mpsc::channel::<(String, u32)>(1000);
+
+    let (yara_memory_tx, yara_memory_rx) =
+        if cfg.scanner.yara_enabled && cfg.scanner.yara_memory_enabled {
+            let capacity = cfg.scanner.yara_memory_queue_capacity.max(1);
+            let (tx, rx) = mpsc::channel::<YaraMemoryJob>(capacity);
+            (Some(tx), Some(rx))
+        } else {
+            (None, None)
+        };
+    let yara_worker_handle = runtime_yara::spawn_yara_file_worker(
+        Arc::clone(&detectors),
+        alert_sink.clone(),
+        response_engine.clone(),
+        cfg.alerts.match_debug,
+        yara_rx,
+        yara_allowlist_paths.clone(),
+        Platform::MacOS,
+        "esf",
+    );
+
+    // Spawn optional YARA memory scanning worker (macOS).
+    let yara_memory_worker_handle = if let Some(mem_rx) = yara_memory_rx {
+        let mem_cfg = MemoryScanConfig {
+            max_process_bytes: (cfg.scanner.yara_memory_max_process_mb * 1024 * 1024) as usize,
+            max_region_bytes: (cfg.scanner.yara_memory_max_region_mb * 1024 * 1024) as usize,
+            include_private: cfg.scanner.yara_memory_include_private,
+            include_image: cfg.scanner.yara_memory_include_image,
+            include_mapped: cfg.scanner.yara_memory_include_mapped,
+            delay_ms: cfg.scanner.yara_memory_delay_ms,
+        };
+        Some(runtime_yara::spawn_yara_memory_worker(
+            Arc::clone(&detectors),
+            alert_sink.clone(),
+            response_engine.clone(),
+            mem_cfg,
+            cfg.alerts.match_debug,
+            mem_rx,
+            Platform::MacOS,
+            "yara-memory",
+        ))
+    } else {
+        None
+    };
+
+    // 10. IOC hash background worker
+    let (ioc_hash_tx, mut ioc_hash_worker_handle) = if ioc_engine.is_enabled() {
+        let (hash_tx, hash_rx) = mpsc::channel::<(String, u32)>(1000);
+        let handle = runtime_ioc::spawn_ioc_hash_worker(
+            Arc::clone(&detectors),
+            alert_sink.clone(),
+            response_engine.clone(),
+            hash_rx,
+            Platform::MacOS,
+            "esf",
+        );
+        (Some(hash_tx), Some(handle))
+    } else {
+        (None, None)
+    };
+
+    // 11. Normalizer
+    let normalizer = Arc::new(Normalizer::new(
+        Arc::clone(&process_cache),
+        Arc::clone(&sid_cache),
+        Arc::clone(&dns_cache),
+        Arc::clone(&connection_aggregator),
+        cfg.network.aggregation_enabled,
+    ));
+
+    // 12. Detection handlers + router
+    let sigma_handler = SigmaDetectionHandler {
+        normalizer: Arc::clone(&normalizer),
+        detectors: Arc::clone(&detectors),
+        ioc_hash_tx,
+        alert_sink: alert_sink.clone(),
+        response_engine: response_engine.clone(),
+    };
+    let yara_handler = YaraEventHandler {
+        tx: yara_tx,
+        memory_tx: yara_memory_tx,
+        allowlist_paths: yara_allowlist_paths,
+    };
+    let mut router_inner = SensorEventRouter::new();
+    router_inner.register_handler(Box::new(sigma_handler));
+    router_inner.register_handler(Box::new(yara_handler));
+    let router = Arc::new(router_inner);
+
+    // 13. macOS Endpoint Security sensor
+    let sensor = Arc::new(EsfSensor::new());
+
+    info!("Starting macOS sensor...");
+    info!("Press Ctrl+C to stop gracefully");
+
+    let (sensor_tx, mut sensor_rx) = mpsc::channel::<SensorEvent>(8192);
+    let router_for_worker = Arc::clone(&router);
+    let sensor_worker_handle = tokio::task::spawn_blocking(move || {
+        while let Some(event) = sensor_rx.blocking_recv() {
+            router_for_worker.route_event(&event);
+        }
+    });
+
+    let sensor_clone = Arc::clone(&sensor);
+    if let Err(e) = sensor_clone.start(sensor_tx) {
+        error!("macOS sensor failed to start: {:#}", e);
+        return Err(e);
+    }
+
+    // 14. Wait for Ctrl+C
+    match tokio::signal::ctrl_c().await {
+        Ok(()) => info!("Received Ctrl+C, shutting down"),
+        Err(e) => error!("Failed to listen for Ctrl+C: {}", e),
+    }
+    sensor.shutdown();
+
+    // Drain workers
+    drop(router);
+    drop(response_engine);
+    let _ = sensor_worker_handle.await;
+    let _ = yara_worker_handle.await;
+    if let Some(h) = yara_memory_worker_handle {
+        let _ = h.await;
+    }
+    if let Some(h) = ioc_hash_worker_handle.take() {
+        let _ = h.await;
+    }
+    if let Some(h) = reload_poller_handle.take() {
+        h.abort();
+        let _ = h.await;
+    }
+    drop(reload_tx.take());
+    if let Some(h) = reload_worker_handle.take() {
+        let _ = h.await;
+    }
+    let _ = response_worker_handle.await;
+
+    info!("Shutdown complete");
+    Ok(())
+}
